lldp: attach remote TTL to port instead of chassis
When receiving a 0 TTL, we should trigger expiry of the remote port, not
the remote chassis. Therefore, TTL is moved to port, except for the
local chassis (we need it even when we don't have a port). So, for
remote ports, TTL is on the port. For local ports, it is on the
chassis. This fixes two cases:

 - a shutdown LLDPU should only expire one local port
 - when a port is refreshed, another port of the same chassis shouldn't
   be refreshed