daemon: don't enable ProtectSystem by default
If the chroot is in `/usr` (like `/usr/local/var/run/lldpd` which is
the default), neither systemd nor lldpd will be able to create and
write to it. This may be solved with `ReadWritePaths` (unsure if it
would create the directory), but this doesn't exist in older versions
of systemd.

Just comment the directive to let people know it exists and should
work in most cases.

Fix #308