daemon: annotate "daemonisation" to help Coverity
When daemonizing, we need to use /dev/null for stdin, stdout, and
stderr. If one of these file descriptors happen to be already closed,
we need to close the new file descriptor only if > 2. This is
confusing confusing for Coverity, annotate it correctly.

Not all locations are annotated because not all of them are detected,
for some reason.