daemon: fix some use of "if defined"
`#if defined HOST_OS_FREEBSD || HOST_OS_NETBSD` is not the same as
`#if defined HOST_OS_FREEBSD || defined HOST_OS_NETBSD`. This doesn't
really matter in our case, but for consistency, fix these cases. Also,
don't use parentheses as they are not useful in our case (or we should
put them everywhere).