build: ship universal binaries for macOS
This mostly reverts commit 737afb3517d822ae50722510993acbd097bc5aca.
We have to build in an OOT directory to make it work. This also
requires to bump minimum SDK to 11.1 (for ARM64 mostly but it also fix
an issue with strtonum() which is detected as available but really is
since 11.0 only).

This does not currently work as the host has to be ARM64 (to be able
to run both binaries). Otherwise, we would have to cross-compile. See
https://github.com/actions/runner-images/issues/2187 for availability.